[namespace] Move Find annotations to serenity_bdd namespace
Create deprecated versions in thucydides namespace but with 2 on name to ensure
caught all changes, and returning objects of correct classes.

Also kept deprecated versions of tests to ensure old versions still work correctly
