+++ core/src/test/java/sample/page/TestPage.java
@@ -1,7 +1,7 @@
 package sample.page;
 
+import net.serenity_bdd.core.annotations.findby.FindBy;
 import net.thucydides.core.annotations.DefaultUrl;
-import net.thucydides.core.annotations.findby.FindBy;
 import net.thucydides.core.pages.PageObject;
 import org.openqa.selenium.WebDriver;
 import sample.elements.WebElementFacadeInput;
