+++ core/src/test/java/sample/page/WithBadImplementerPage.java
@@ -1,6 +1,6 @@
 package sample.page;
 
-import net.thucydides.core.annotations.findby.FindBy;
+import net.serenity_bdd.core.annotations.findby.FindBy;
 import net.thucydides.core.pages.PageObject;
 import org.openqa.selenium.WebDriver;
 import sample.elements.HasBadImplementer;
